chore: regenerate minified pixel
